requirements.txt
sqlalchemy
psycopg2-binary
chromadb
ollama
openai
python-dotenv
pandas
sentence-transformers
tabulate
PyYAML